+++ core-ui/src/components/CustomResources/CustomResourcesByGroup.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import LuigiClient from '@luigi-project/client';
 import { Trans, useTranslation } from 'react-i18next';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/Link/ExternalLink';
 import { Link as FdLink } from 'fundamental-react';
 import pluralize from 'pluralize';
 import { GroupingListPage } from './GroupingListPage';
@@ -17,7 +17,7 @@ export default function CustomResourcesByGroup({ namespace }) {
 
   const description = (
     <Trans i18nKey="custom-resources.description">
-      <Link
+      <ExternalLink
         className="fd-link"
         url="https://kubernetes.io/docs/concepts/extend-kubernetes/api-extension/custom-resources/"
       />
