+++ core-ui/src/resources/AddonsConfigurations/GenericAddonsConfigurationList.js
@@ -3,7 +3,7 @@ import { useTranslation, Trans } from 'react-i18next';
 
 import { ResourcesList } from 'shared/components/ResourcesList/ResourcesList';
 import { ResourceStatus } from 'shared/components/ResourceStatus/ResourceStatus';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/Link/ExternalLink';
 
 export function GenericAddonsConfigurationList({
   descriptionKey,
@@ -21,7 +21,7 @@ export function GenericAddonsConfigurationList({
 
   const description = (
     <Trans i18nKey={descriptionKey}>
-      <Link className="fd-link" url={documentationLink} />
+      <ExternalLink className="fd-link" url={documentationLink} />
     </Trans>
   );
