+++ core-ui/src/resources/ConfigMaps/ConfigMapList.js
@@ -3,7 +3,7 @@ import { useTranslation, Trans } from 'react-i18next';
 
 import { ResourcesList } from 'shared/components/ResourcesList/ResourcesList';
 import { ControlledByKind } from 'shared/components/ControlledBy/ControlledBy';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/Link/ExternalLink';
 
 import { ConfigMapCreate } from './ConfigMapCreate';
 
@@ -21,7 +21,7 @@ export function ConfigMapList(props) {
 
   const description = (
     <Trans i18nKey="config-maps.description">
-      <Link
+      <ExternalLink
         className="fd-link"
         url="https://kubernetes.io/docs/concepts/configuration/configmap/"
       />
