+++ core-ui/src/resources/DestinationRules/DestinationRuleList.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { useTranslation, Trans } from 'react-i18next';
 
 import { ResourcesList } from 'shared/components/ResourcesList/ResourcesList';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/Link/ExternalLink';
 
 import { DestinationRuleCreate } from './DestinationRuleCreate';
 
@@ -18,7 +18,7 @@ export function DestinationRuleList(props) {
 
   const description = (
     <Trans i18nKey="destination-rules.description">
-      <Link
+      <ExternalLink
         className="fd-link"
         url="https://istio.io/latest/docs/reference/config/networking/destination-rule/"
       />
