+++ core-ui/src/resources/DnsEntries/DnsEntryList.js
@@ -3,7 +3,7 @@ import { useTranslation, Trans } from 'react-i18next';
 
 import { ResourcesList } from 'shared/components/ResourcesList/ResourcesList';
 import { ResourceStatus } from 'shared/components/ResourceStatus/ResourceStatus';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/Link/ExternalLink';
 
 import { DnsEntryCreate } from './DnsEntryCreate';
 
@@ -24,7 +24,7 @@ export function DnsEntryList(params) {
 
   const description = (
     <Trans i18nKey="dnsentries.description">
-      <Link
+      <ExternalLink
         className="fd-link"
         url="https://kyma-project.io/docs/kyma/latest/03-tutorials/00-api-exposure/apix-01-own-domain"
       />
