+++ core-ui/src/resources/Ingresses/IngressList.js
@@ -3,7 +3,7 @@ import { useTranslation, Trans } from 'react-i18next';
 
 import { ResourcesList } from 'shared/components/ResourcesList/ResourcesList';
 import { EMPTY_TEXT_PLACEHOLDER } from 'shared/constants';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/Link/ExternalLink';
 
 import { IngressCreate } from './IngressCreate';
 
@@ -29,7 +29,7 @@ export function IngressList(props) {
 
   const description = (
     <Trans i18nKey="ingresses.description">
-      <Link
+      <ExternalLink
         className="fd-link"
         url="https://kubernetes.io/docs/concepts/services-networking/ingress/"
       />
