+++ core-ui/src/resources/Issuers/IssuerList.js
@@ -3,7 +3,7 @@ import { useTranslation, Trans } from 'react-i18next';
 
 import { ResourcesList } from 'shared/components/ResourcesList/ResourcesList';
 import { ResourceStatus } from 'shared/components/ResourceStatus/ResourceStatus';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/Link/ExternalLink';
 
 import { IssuerCreate } from './IssuerCreate';
 
@@ -29,7 +29,7 @@ export function IssuerList(props) {
 
   const description = (
     <Trans i18nKey="issuers.description">
-      <Link
+      <ExternalLink
         className="fd-link"
         url="https://cert-manager.io/docs/concepts/issuer/"
       />
