+++ core-ui/src/resources/Jobs/JobList.js
@@ -3,7 +3,7 @@ import { useTranslation, Trans } from 'react-i18next';
 
 import { ResourcesList } from 'shared/components/ResourcesList/ResourcesList';
 import { ControlledBy } from 'shared/components/ControlledBy/ControlledBy';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/Link/ExternalLink';
 
 import { JobCompletions } from './JobCompletions';
 import { JobCreate } from './JobCreate';
@@ -25,7 +25,7 @@ export const JobList = props => {
 
   const description = (
     <Trans i18nKey="jobs.description">
-      <Link
+      <ExternalLink
         className="fd-link"
         url="https://kubernetes.io/docs/concepts/workloads/controllers/job/"
       />
