+++ core-ui/src/resources/Secrets/SecretList.js
@@ -3,7 +3,7 @@ import { useTranslation, Trans } from 'react-i18next';
 
 import { ResourcesList } from 'shared/components/ResourcesList/ResourcesList';
 import { ControlledByKind } from 'shared/components/ControlledBy/ControlledBy';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/Link/ExternalLink';
 
 import { SecretCreate } from './SecretCreate';
 
@@ -27,7 +27,7 @@ export function SecretList(props) {
 
   const description = (
     <Trans i18nKey="secrets.description">
-      <Link
+      <ExternalLink
         className="fd-link"
         url="https://kubernetes.io/docs/concepts/configuration/secret/"
       />
