+++ core-ui/src/resources/ServiceInstances/ServiceInstanceList.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { BTPResourceStatus } from 'shared/components/BTPResourceStatus';
 import { ResourcesList } from 'shared/components/ResourcesList/ResourcesList';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/Link/ExternalLink';
 import { Trans } from 'react-i18next';
 import { ServiceInstanceCreate } from './ServiceInstanceCreate';
 
@@ -35,7 +35,7 @@ export function ServiceInstanceList(props) {
 
   const description = (
     <Trans i18nKey="btp-instances.description">
-      <Link
+      <ExternalLink
         className="fd-link"
         url="https://github.com/SAP/sap-btp-service-operator#step-1-create-a-service-instance"
       />
