+++ core-ui/src/resources/Services/ServiceList.js
@@ -3,7 +3,7 @@ import { useTranslation, Trans } from 'react-i18next';
 
 import { ResourcesList } from 'shared/components/ResourcesList/ResourcesList';
 import { ControlledByKind } from 'shared/components/ControlledBy/ControlledBy';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/Link/ExternalLink';
 
 import { ServiceCreate } from './ServiceCreate';
 
@@ -65,7 +65,7 @@ export function ServiceList(props) {
 
   const description = (
     <Trans i18nKey="services.description">
-      <Link
+      <ExternalLink
         className="fd-link"
         url="https://kubernetes.io/docs/concepts/services-networking/service/"
       />
