+++ core-ui/src/resources/StatefulSets/StatefulSetList.js
@@ -3,7 +3,7 @@ import { useTranslation, Trans } from 'react-i18next';
 
 import { ResourcesList } from 'shared/components/ResourcesList/ResourcesList';
 import { ControlledByKind } from 'shared/components/ControlledBy/ControlledBy';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/Link/ExternalLink';
 
 import { useRestartAction } from 'shared/hooks/useRestartResource';
 import { StatefulSetCreate } from './StatefulSetCreate';
@@ -28,7 +28,7 @@ export function StatefulSetList(props) {
 
   const description = (
     <Trans i18nKey="stateful-sets.description">
-      <Link
+      <ExternalLink
         className="fd-link"
         url="https://kubernetes.io/docs/concepts/workloads/controllers/statefulset/"
       />
