+++ core-ui/src/resources/Subscriptions/SubscriptionList.js
@@ -3,7 +3,7 @@ import { useTranslation, Trans } from 'react-i18next';
 
 import { ResourcesList } from 'shared/components/ResourcesList/ResourcesList';
 import { EMPTY_TEXT_PLACEHOLDER } from 'shared/constants';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/Link/ExternalLink';
 import { SubscriptionConditionStatus } from 'shared/components/SubscriptionConditionStatus';
 
 import { SubscriptionCreate } from './SubscriptionCreate';
@@ -54,7 +54,7 @@ export function SubscriptionList(props) {
 
   const description = (
     <Trans i18nKey="subscriptions.description">
-      <Link
+      <ExternalLink
         className="fd-link"
         url="https://kyma-project.io/docs/kyma/latest/05-technical-reference/00-custom-resources/evnt-01-subscription/#documentation-content"
       />
