+++ core-ui/src/resources/VirtualServices/VirtualServiceList.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { ResourcesList } from 'shared/components/ResourcesList/ResourcesList';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/Link/ExternalLink';
 import { VirtualServiceCreate } from './VirtualServiceCreate';
 import { Trans } from 'react-i18next';
 
@@ -22,7 +22,7 @@ export function VirtualServiceList(props) {
 
   const description = (
     <Trans i18nKey="virtualservices.description">
-      <Link
+      <ExternalLink
         className="fd-link"
         url="https://istio.io/latest/docs/reference/config/networking/virtual-service/"
       />
