+++ core-ui/src/shared/components/Link/CopiableLink.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { CopiableText } from 'shared/components/CopiableText/CopiableText';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/Link/ExternalLink';
 
 export const CopiableLink = props => {
   return (
     <CopiableText textToCopy={props.url} compact={true} i18n={props.i18n}>
-      <Link className="fd-link" {...props} />
+      <ExternalLink className="fd-link" {...props} />
     </CopiableText>
   );
 };
 
-CopiableLink.propTypes = Link.propTypes;
+CopiableLink.propTypes = ExternalLink.propTypes;
